back-out configuration-related changes
