+++ linkerd/app/inbound/src/endpoint.rs
@@ -122,7 +122,7 @@ impl<A> router::Recognize<http::Request<A>> for RecognizeEndpoint {
     fn recognize(&self, req: &http::Request<A>) -> Option<Self::Target> {
         let src = req.extensions().get::<tls::accept::Meta>();
         debug!("inbound endpoint: src={:?}", src);
-        let addr = src.and_then(|m| m.addrs.target_addr_if_not_local())?;
+        let addr = src.and_then(|s| s.addrs.target_addr_if_not_local())?;
 
         let tls_client_id = src
             .map(|s| s.peer_identity.clone())
