+++ linkerd/app/inbound/src/set_remote_ip_on_req.rs
@@ -1,5 +1,5 @@
 //! Adds `l5d-remote-ip` headers to http::Requests derived from the
-//! `remote` of a `tls::accept::Meta`.
+//! `remote` of a `Source`.
 
 use bytes::Bytes;
 use http::header::HeaderValue;
