+++ linkerd/app/integration/tests/tap.rs
@@ -17,7 +17,7 @@ fn tap_enabled_when_identity_enabled() {
         .identity(identity_env.service().run())
         .run_with_test_env(identity_env.env);
 
-    assert!(proxy.tap.is_some())
+    assert!(proxy.control.is_some())
 }
 
 #[test]
@@ -25,18 +25,18 @@ fn tap_disabled_when_identity_disabled() {
     let _ = trace_init();
     let proxy = proxy::new().disable_identity().run();
 
-    assert!(proxy.tap.is_none())
+    assert!(proxy.control.is_none())
 }
 
 #[test]
 fn can_disable_tap() {
     let _ = trace_init();
-    let mut env = TestEnv::new();
+    let mut env = app::config::TestEnv::new();
     env.put(app::config::ENV_TAP_DISABLED, "true".to_owned());
 
     let proxy = proxy::new().run_with_test_env(env);
 
-    assert!(proxy.tap.is_none())
+    assert!(proxy.control.is_none())
 }
 
 #[test]
@@ -58,7 +58,7 @@ fn rejects_incorrect_identity_when_identity_is_expected() {
         .run_with_test_env(expected_identity_env);
 
     let tap_proxy = proxy::new()
-        .outbound_ip(in_proxy.tap.unwrap())
+        .outbound_ip(in_proxy.control.unwrap())
         .identity(identity_env.service().run())
         .run_with_test_env(identity_env.env.clone());
 
@@ -112,7 +112,7 @@ fn inbound_http1() {
 
     // Run client proxy with client proxy identity service.
     let client_proxy = proxy::new()
-        .outbound_ip(srv_proxy.tap.unwrap())
+        .outbound_ip(srv_proxy.control.unwrap())
         .identity(client_proxy_identity.service().run())
         .run_with_test_env(client_proxy_identity.env);
 
@@ -188,7 +188,7 @@ fn grpc_headers_end() {
 
     // Run client proxy with client proxy identity service.
     let client_proxy = proxy::new()
-        .outbound_ip(srv_proxy.tap.unwrap())
+        .outbound_ip(srv_proxy.control.unwrap())
         .identity(client_proxy_identity.service().run())
         .run_with_test_env(client_proxy_identity.env);
