+++ linkerd/app/outbound/src/add_remote_ip_on_rsp.rs
@@ -1,5 +1,5 @@
 //! Adds `l5d-remote-ip` headers to http::Responses derived from the
-//! `remote` of a `tls::accept::Meta`.
+//! `remote` of a `Source`.
 
 use super::Endpoint;
 use bytes::Bytes;
