+++ linkerd/app/outbound/src/lib.rs
@@ -6,15 +6,15 @@
 #![deny(warnings, rust_2018_idioms)]
 
 use linkerd2_app_core::{
-    self as core, classify,
-    config::ProxyConfig,
+    classify,
+    config::Config,
     dns, drain,
     dst::DstAddr,
     errors, http_request_authority_addr, http_request_host_addr,
     http_request_l5d_override_dst_addr, http_request_orig_dst_addr, identity,
     proxy::{
         self,
-        core::{resolve::Resolve, Listen},
+        core::resolve::Resolve,
         discover,
         http::{
             balance, canonicalize, client, fallback, header_from_target, insert,
@@ -26,7 +26,7 @@ use linkerd2_app_core::{
     reconnect, serve,
     spans::SpanConverter,
     svc, trace, trace_context,
-    transport::{self, connect, tls, OrigDstAddr, SysOrigDstAddr},
+    transport::{self, connect, tls, OrigDstAddr},
     Addr, Conditional, DispatchDeadline, CANONICAL_DST_HEADER, DST_OVERRIDE_HEADER, L5D_CLIENT_ID,
     L5D_REMOTE_IP, L5D_REQUIRE_ID, L5D_SERVER_ID,
 };
@@ -52,41 +52,23 @@ pub use self::resolve::resolve;
 const EWMA_DEFAULT_RTT: Duration = Duration::from_millis(30);
 const EWMA_DECAY: Duration = Duration::from_secs(10);
 
-#[derive(Clone, Debug)]
-pub struct Config<O: OrigDstAddr = SysOrigDstAddr> {
-    pub proxy: ProxyConfig<O>,
-    pub canonicalize_timeout: Duration,
-}
-
-impl<O: OrigDstAddr> Config<O> {
-    pub fn with_orig_dst_addrs_from<P: OrigDstAddr>(self, orig_dst_addrs: P) -> Config<P> {
-        Config {
-            proxy: self.proxy.with_orig_dst_addrs_from(orig_dst_addrs.clone()),
-            canonicalize_timeout: self.canonicalize_timeout,
-        }
-    }
-}
-
-pub fn spawn<O, L, R, P>(
-    listen: L,
-    config: Config<O>,
+pub fn spawn<A, R, P>(
+    config: &Config,
     local_identity: tls::Conditional<identity::Local>,
+    listen: transport::Listen<A>,
     resolve: R,
     dns_resolver: dns::Resolver,
-    profiles_client: core::profiles::Client<P>,
-    profile_suffixes: Vec<dns::Suffix>,
-    tap_layer: core::tap::Layer,
-    span_sink: Option<mpsc::Sender<oc::Span>>,
+    profiles_client: linkerd2_app_core::profiles::Client<P>,
+    tap_layer: linkerd2_app_core::tap::Layer,
     handle_time: http_metrics::handle_time::Scope,
-    endpoint_http_metrics: core::HttpEndpointMetricsRegistry,
-    route_http_metrics: core::HttpRouteMetricsRegistry,
-    retry_http_metrics: core::HttpRouteMetricsRegistry,
-    transport_metrics: transport::MetricsRegistry,
+    endpoint_http_metrics: linkerd2_app_core::HttpEndpointMetricsRegistry,
+    route_http_metrics: linkerd2_app_core::HttpRouteMetricsRegistry,
+    retry_http_metrics: linkerd2_app_core::HttpRouteMetricsRegistry,
+    transport_metrics: linkerd2_app_core::transport::MetricsRegistry,
+    span_sink: Option<mpsc::Sender<oc::Span>>,
     drain: drain::Watch,
 ) where
-    O: transport::OrigDstAddr + Send + 'static,
-    L: Listen<Connection = transport::listen::Connection> + Send + 'static,
-    L::Error: std::error::Error + Send + 'static,
+    A: OrigDstAddr + Send + 'static,
     R: Resolve<DstAddr, Endpoint = Endpoint> + Clone + Send + Sync + 'static,
     R::Future: Send,
     R::Resolution: Send,
@@ -95,20 +77,21 @@ pub fn spawn<O, L, R, P>(
     <P::ResponseBody as grpc::Body>::Data: Send,
     P::Future: Send,
 {
-    let capacity = config.proxy.router_capacity;
-    let max_idle_age = config.proxy.router_max_idle_age;
-    let max_in_flight = config.proxy.server.buffer.max_in_flight;
-    let dispatch_timeout = config.proxy.server.buffer.dispatch_timeout;
-    let canonicalize_timeout = config.canonicalize_timeout;
+    let capacity = config.outbound_router_capacity;
+    let max_idle_age = config.outbound_router_max_idle_age;
+    let max_in_flight = config.outbound_max_requests_in_flight;
+    let profile_suffixes = config.destination_profile_suffixes.clone();
+    let canonicalize_timeout = config.dns_canonicalize_timeout;
+    let dispatch_timeout = config.outbound_dispatch_timeout;
 
     let mut trace_labels = HashMap::new();
     trace_labels.insert("direction".to_string(), "outbound".to_string());
 
     // Establishes connections to remote peers (for both TCP
     // forwarding and HTTP proxying).
-    let connect = svc::stack(connect::svc(config.proxy.connect.keepalive))
+    let connect = svc::stack(connect::svc(config.outbound_connect_keepalive))
         .push(tls::client::layer(local_identity))
-        .push_timeout(config.proxy.connect.timeout)
+        .push_timeout(config.outbound_connect_timeout)
         .push(transport_metrics.layer_connect(TransportLabels));
 
     let trace_context_layer = trace_context::layer(
@@ -119,9 +102,9 @@ pub fn spawn<O, L, R, P>(
     // Instantiates an HTTP client for for a `client::Config`
     let client_stack = connect
         .clone()
-        .push(client::layer(config.proxy.connect.h2_settings))
+        .push(client::layer(config.h2_settings))
         .push(reconnect::layer({
-            let backoff = config.proxy.connect.backoff.clone();
+            let backoff = config.outbound_connect_backoff.clone();
             move |_| Ok(backoff.stream())
         }))
         .push(trace_context_layer)
@@ -154,7 +137,7 @@ pub fn spawn<O, L, R, P>(
         ))
         .push(require_identity_on_endpoint::layer())
         .push(trace::layer(|endpoint: &Endpoint| {
-            info_span!("endpoint", peer.addr = %endpoint.addr, peer.id = ?endpoint.identity)
+            info_span!("endpoint", ?endpoint)
         }));
 
     // A per-`dst::Route` layer that uses profile data to configure
@@ -270,7 +253,7 @@ pub fn spawn<O, L, R, P>(
     //
     // 4. If the request has an HTTP/1 Host header, it is used.
     //
-    // 5. Finally, if the tls::accept::Meta had an SO_ORIGINAL_DST, this TCP
+    // 5. Finally, if the Source had an SO_ORIGINAL_DST, this TCP
     // address is used.
     let addr_router = addr_stack
         .push(strip_header::request::layer(L5D_CLIENT_ID))
@@ -304,8 +287,8 @@ pub fn spawn<O, L, R, P>(
     let trace_context_layer = trace_context::layer(
         span_sink.map(|span_sink| SpanConverter::server(span_sink, trace_labels)),
     );
-    // Instantiates an HTTP service for each `tls::accept::Meta` using the
-    // shared `addr_router`. The `tls::accept::Meta` is stored in the request's
+    // Instantiates an HTTP service for each `Source` using the
+    // shared `addr_router`. The `Source` is stored in the request's
     // extensions so that it can be used by the `addr_router`.
     let server_stack = svc::stack(svc::Shared::new(admission_control))
         .push(insert::layer(move || {
@@ -319,8 +302,7 @@ pub fn spawn<O, L, R, P>(
         .push(trace_context_layer)
         .push(handle_time.layer());
 
-    let skip_ports = config.proxy.disable_protocol_detection_for_ports;
-
+    let skip_ports = std::sync::Arc::new(config.outbound_ports_disable_protocol_detection.clone());
     let proxy = Server::new(
         TransportLabels,
         transport_metrics,
@@ -328,7 +310,7 @@ pub fn spawn<O, L, R, P>(
             .push(svc::map_target::layer(Endpoint::from))
             .into_inner(),
         server_stack,
-        config.proxy.server.h2_settings,
+        config.h2_settings,
         drain.clone(),
         skip_ports.clone(),
     );
