+++ linkerd/app/src/lib.rs
@@ -2,27 +2,28 @@
 
 #![deny(warnings, rust_2018_idioms)]
 
-pub mod config;
-
-use self::config::*;
 use futures::{self, future, Future};
-pub use linkerd2_app_core as core;
 use linkerd2_app_core::{
     admin::{Admin, Readiness},
     classify::{self, Class},
-    config::*,
+    config::Config,
     control, dns, drain, handle_time, identity,
     metric_labels::{ControlLabels, EndpointLabels, RouteLabels},
     metrics::FmtMetrics,
     opencensus::{self, SpanExporter},
     profiles::Client as ProfilesClient,
-    proxy::{self, api_resolve, core::listen::Listen, http::metrics as http_metrics},
+    proxy::{
+        self, api_resolve,
+        core::listen::{Bind as _Bind, Listen as _Listen},
+        http::metrics as http_metrics,
+    },
     reconnect, serve,
     svc::{self, LayerExt},
     tap, telemetry, trace,
-    transport::{self, connect, listen, tls, OrigDstAddr, SysOrigDstAddr},
-    Conditional, Error,
+    transport::{self, connect, listen, tls, OrigDstAddr},
+    Conditional,
 };
+pub use linkerd2_app_core::{init, transport::SysOrigDstAddr};
 use linkerd2_app_inbound as inbound;
 use linkerd2_app_outbound as outbound;
 use opencensus_proto::agent::common::v1 as oc;
@@ -46,121 +47,105 @@ use tracing_futures::Instrument;
 ///
 /// The private listener routes requests to service-discovery-aware load-balancer.
 ///
-#[derive(Clone, Debug)]
-pub struct Main<O: OrigDstAddr = SysOrigDstAddr> {
-    config: Config<O>,
-    trace_handle: trace::LevelHandle,
+pub struct Main<A: OrigDstAddr> {
+    proxy_parts: ProxyParts<A>,
 }
 
-pub struct Bound<O: OrigDstAddr = SysOrigDstAddr> {
+struct ProxyParts<A: OrigDstAddr> {
+    config: Config,
+
     start_time: SystemTime,
-    trace_handle: trace::LevelHandle,
+    trace_level: trace::LevelHandle,
 
-    dns: (dns::Resolver, dns::Task),
-    dst: DestinationConfig,
-    identity: tls::Conditional<IdentityConfig>,
-    trace_collector: Option<TraceCollectorConfig>,
+    admin_listener: listen::Listen,
+    control_listener: Option<(listen::Listen, identity::Name)>,
 
-    admin: (listen::Listen, AdminConfig),
-    tap: Option<(listen::Listen, TapConfig)>,
-    inbound: (listen::Listen<O>, ProxyConfig<O>),
-    outbound: (listen::Listen<O>, outbound::Config<O>),
+    inbound_listener: listen::Listen<A>,
+    outbound_listener: listen::Listen<A>,
 }
 
-impl Main {
-    pub fn try_from_env() -> Result<Self, Error> {
-        let config = Config::try_from_env()?;
-        let trace_handle = core::trace::init()?;
-        Ok(Self::new(config, trace_handle))
-    }
-
-    pub fn new(config: Config, trace_handle: trace::LevelHandle) -> Self {
-        Self {
-            config,
-            trace_handle,
-        }
-    }
-}
-
-impl<O: OrigDstAddr> Main<O> {
-    pub fn with_orig_dst_addrs_from<P: OrigDstAddr>(self, orig_dst_addrs: P) -> Main<P> {
-        Main {
-            config: self.config.with_orig_dst_addrs_from(orig_dst_addrs),
-            trace_handle: self.trace_handle,
-        }
-    }
-
-    pub fn bind(self) -> Result<Bound<O>, Error> {
-        use proxy::core::listen::Bind;
-
+impl<A: OrigDstAddr> Main<A> {
+    pub fn new(config: Config, trace_level: trace::LevelHandle, orig_dsts: A) -> Self {
         let start_time = SystemTime::now();
 
-        let tap = if let Some(config) = self.config.tap {
-            let listen = config.server.bind.clone().bind().map_err(Error::from)?;
-            Some((listen, config))
-        } else {
-            None
-        };
+        let control_listener = config.control_listener.as_ref().map(|cl| {
+            let listener = listen::Bind::new(cl.listener.addr, config.inbound_accept_keepalive)
+                .bind()
+                .expect("tap listener bind");
 
-        let dns = dns::Resolver::from_system_config_with(&self.config.dns)
-            .expect("invalid DNS configuration");
+            (listener, cl.tap_svc_name.clone())
+        });
 
-        let admin = (
-            self.config.admin.server.bind.clone().bind()?,
-            self.config.admin,
-        );
-        let inbound = (
-            self.config.inbound.server.bind.clone().bind()?,
-            self.config.inbound,
-        );
-        let outbound = (
-            self.config.outbound.proxy.server.bind.clone().bind()?,
-            self.config.outbound,
-        );
+        let admin_listener =
+            listen::Bind::new(config.admin_listener.addr, config.inbound_accept_keepalive)
+                .bind()
+                .expect("tap listener bind");
 
-        Ok(Bound {
-            start_time,
-            dns,
-            tap,
-            admin,
-            inbound,
-            outbound,
-            dst: self.config.destination,
-            identity: self.config.identity,
-            trace_handle: self.trace_handle,
-            trace_collector: self.config.trace_collector,
-        })
+        let inbound_listener = listen::Bind::new(
+            config.inbound_listener.addr,
+            config.inbound_accept_keepalive,
+        )
+        .with_orig_dst_addr(orig_dsts.clone())
+        .bind()
+        .expect("inbound listener bind");
+
+        let outbound_listener = listen::Bind::new(
+            config.outbound_listener.addr,
+            config.outbound_accept_keepalive,
+        )
+        .with_orig_dst_addr(orig_dsts)
+        .bind()
+        .expect("outbound listener bind");
+
+        Self {
+            proxy_parts: ProxyParts {
+                config,
+                start_time,
+                trace_level,
+                inbound_listener,
+                outbound_listener,
+                control_listener,
+                admin_listener,
+            },
+        }
     }
 }
 
-impl<O> Bound<O>
+impl<A> Main<A>
 where
-    O: OrigDstAddr + Send + Sync + 'static,
+    A: OrigDstAddr + Clone + Send + 'static,
 {
-    pub fn tap_addr(&self) -> Option<SocketAddr> {
-        self.tap.as_ref().map(|l| l.0.listen_addr().clone())
+    pub fn control_addr(&self) -> Option<SocketAddr> {
+        self.proxy_parts
+            .control_listener
+            .as_ref()
+            .map(|l| l.0.listen_addr().clone())
     }
 
     pub fn inbound_addr(&self) -> SocketAddr {
-        self.inbound.0.listen_addr()
+        self.proxy_parts.inbound_listener.listen_addr()
     }
 
     pub fn outbound_addr(&self) -> SocketAddr {
-        self.outbound.0.listen_addr()
+        self.proxy_parts.outbound_listener.listen_addr()
     }
 
     pub fn metrics_addr(&self) -> SocketAddr {
-        self.admin.0.listen_addr()
+        self.proxy_parts.admin_listener.listen_addr()
     }
 
-    pub fn run_until(self, shutdown_signal: impl Future<Item = (), Error = ()> + Send + 'static) {
+    pub fn run_until<F>(self, shutdown_signal: F)
+    where
+        F: Future<Item = (), Error = ()> + Send + 'static,
+    {
+        let Main { proxy_parts } = self;
         Runtime::new()
             .expect("runtime")
             .block_on(futures::lazy(move || {
                 let (drain_tx, drain_rx) = drain::channel();
 
-                trace!("spawning proxy");
-                self.spawn_proxy_task(drain_rx);
+                proxy_parts.build_proxy_task(drain_rx);
+                trace!("main task spawned");
 
                 shutdown_signal.and_then(move |()| {
                     debug!("shutdown signaled");
@@ -169,59 +154,67 @@ where
             }))
             .expect("main");
     }
+}
 
+impl<A> ProxyParts<A>
+where
+    A: OrigDstAddr + Clone + Send + 'static,
+{
     /// This is run inside a `futures::lazy`, so the default Executor is
     /// setup for use in here.
-    fn spawn_proxy_task(self, drain_rx: drain::Watch) {
-        let Bound {
+    fn build_proxy_task(self, drain_rx: drain::Watch) {
+        let ProxyParts {
+            config,
             start_time,
-            dns: (dns_resolver, dns_task),
-            dst,
-            identity,
-            trace_handle,
-            trace_collector,
-            tap,
-            inbound: (inbound_listen, inbound),
-            outbound: (outbound_listen, outbound),
-            admin: (admin_listen, admin),
+            trace_level,
+            control_listener,
+            inbound_listener,
+            outbound_listener,
+            admin_listener,
         } = self;
 
-        info!("destinations from {:?}", dst.control.addr);
-        match identity.as_ref() {
-            Conditional::Some(config) => info!("identity from {:?}", config.control.addr),
+        info!("using destination service at {:?}", config.destination_addr);
+        match config.identity_config.as_ref() {
+            Conditional::Some(config) => info!("using identity service at {:?}", config.svc.addr),
             Conditional::None(reason) => info!("identity is DISABLED: {}", reason),
         }
-        info!("outbound on {}", outbound_listen.listen_addr());
-        info!("inbound on {}", inbound_listen.listen_addr(),);
-        info!("admin on {}", admin_listen.listen_addr(),);
-        match tap.as_ref() {
-            Some((l, _)) => info!("tap on {}", l.listen_addr()),
-            None => info!("tap is DISABLED"),
-        }
+        info!("admin on {}", admin_listener.listen_addr());
+        info!(
+            "tap on {:?}",
+            control_listener.as_ref().map(|l| l.0.listen_addr())
+        );
+        info!("outbound on {:?}", outbound_listener.listen_addr());
+        info!("inbound on {}", inbound_listener.listen_addr());
         info!(
             "protocol detection disabled for inbound ports {:?}",
-            inbound.disable_protocol_detection_for_ports,
+            config.inbound_ports_disable_protocol_detection,
         );
         info!(
             "protocol detection disabled for outbound ports {:?}",
-            outbound.proxy.disable_protocol_detection_for_ports,
+            config.outbound_ports_disable_protocol_detection,
         );
 
+        let (dns_resolver, dns_bg) = dns::Resolver::from_system_config_with(&config)
+            .unwrap_or_else(|e| {
+                // FIXME: DNS configuration should be infallible.
+                panic!("invalid DNS configuration: {:?}", e);
+            });
+
         let (ctl_http_metrics, ctl_http_report) = {
-            let (m, r) = http_metrics::new::<ControlLabels, Class>(admin.metrics_retain_idle);
+            let (m, r) = http_metrics::new::<ControlLabels, Class>(config.metrics_retain_idle);
             (m, r.with_prefix("control"))
         };
 
         let (endpoint_http_metrics, endpoint_http_report) =
-            http_metrics::new::<EndpointLabels, Class>(admin.metrics_retain_idle);
+            http_metrics::new::<EndpointLabels, Class>(config.metrics_retain_idle);
 
         let (route_http_metrics, route_http_report) = {
-            let (m, r) = http_metrics::new::<RouteLabels, Class>(admin.metrics_retain_idle);
+            let (m, r) = http_metrics::new::<RouteLabels, Class>(config.metrics_retain_idle);
             (m, r.with_prefix("route"))
         };
 
         let (retry_http_metrics, retry_http_report) = {
-            let (m, r) = http_metrics::new::<RouteLabels, Class>(admin.metrics_retain_idle);
+            let (m, r) = http_metrics::new::<RouteLabels, Class>(config.metrics_retain_idle);
             (m, r.with_prefix("route_actual"))
         };
 
@@ -245,22 +238,31 @@ where
 
         let mut identity_daemon = None;
         let (readiness, ready_latch) = Readiness::new();
-        let local_identity = match identity {
+        let listen_identity = match config.identity_config.clone() {
             Conditional::None(r) => {
                 ready_latch.release();
                 Conditional::None(r)
             }
-            Conditional::Some(config) => {
-                let (local_identity, crt_store) = identity::Local::new(&config.identity);
-                let svc = svc::stack(connect::svc(config.control.connect.keepalive))
+            Conditional::Some(id_config) => {
+                let (listen_identity, crt_store) = identity::Local::new(&id_config);
+
+                // If the service is on localhost, use the inbound keepalive.
+                // If the service. is remote, use the outbound keepalive.
+                let keepalive = if id_config.svc.addr.is_loopback() {
+                    config.inbound_connect_keepalive
+                } else {
+                    config.outbound_connect_keepalive
+                };
+
+                let svc = svc::stack(connect::svc(keepalive))
                     .push(tls::client::layer(Conditional::Some(
-                        config.identity.trust_anchors.clone(),
+                        id_config.trust_anchors.clone(),
                     )))
-                    .push_timeout(config.control.connect.timeout)
+                    .push_timeout(config.control_connect_timeout)
                     .push(control::client::layer())
                     .push(control::resolve::layer(dns_resolver.clone()))
                     .push(reconnect::layer({
-                        let backoff = config.control.connect.backoff;
+                        let backoff = config.control_backoff.clone();
                         move |_| Ok(backoff.stream())
                     }))
                     .push(http_metrics::layer::<_, classify::Response>(
@@ -269,16 +271,16 @@ where
                     .push(proxy::grpc::req_body_as_payload::layer().per_make())
                     .push(control::add_origin::layer())
                     .push_buffer_pending(
-                        config.control.buffer.max_in_flight,
-                        config.control.buffer.dispatch_timeout,
+                        config.destination_buffer_capacity,
+                        config.control_dispatch_timeout,
                     )
                     .into_inner()
-                    .make(config.control.addr);
+                    .make(id_config.svc.clone());
 
-                identity_daemon = Some(identity::Daemon::new(config.identity, crt_store, svc));
+                identity_daemon = Some(identity::Daemon::new(id_config, crt_store, svc));
 
                 tokio::spawn(
-                    local_identity
+                    listen_identity
                         .clone()
                         .await_crt()
                         .map(move |id| {
@@ -292,39 +294,50 @@ where
                         .instrument(info_span!("await identity")),
                 );
 
-                Conditional::Some(local_identity)
+                Conditional::Some(listen_identity)
             }
         };
 
-        let dst_svc = svc::stack(connect::svc(dst.control.connect.keepalive))
-            .push(tls::client::layer(local_identity.clone()))
-            .push_timeout(dst.control.connect.timeout)
-            .push(control::client::layer())
-            .push(control::resolve::layer(dns_resolver.clone()))
-            .push(reconnect::layer({
-                let backoff = dst.control.connect.backoff;
-                move |_| Ok(backoff.stream())
-            }))
-            .push(http_metrics::layer::<_, classify::Response>(
-                ctl_http_metrics.clone(),
-            ))
-            .push(proxy::grpc::req_body_as_payload::layer().per_make())
-            .push(control::add_origin::layer())
-            .push_buffer_pending(
-                dst.control.buffer.max_in_flight,
-                dst.control.buffer.dispatch_timeout,
-            )
-            .into_inner()
-            .make(dst.control.addr.clone());
+        let dst_svc = {
+            // If the dst_svc is on localhost, use the inbound keepalive.
+            // If the dst_svc is remote, use the outbound keepalive.
+            let keepalive = if config.destination_addr.addr.is_loopback() {
+                config.inbound_connect_keepalive
+            } else {
+                config.outbound_connect_keepalive
+            };
 
-        let resolver = api_resolve::Resolve::new(dst_svc.clone()).with_context_token(&dst.context);
+            svc::stack(connect::svc(keepalive))
+                .push(tls::client::layer(listen_identity.clone()))
+                .push_timeout(config.control_connect_timeout)
+                .push(control::client::layer())
+                .push(control::resolve::layer(dns_resolver.clone()))
+                .push(reconnect::layer({
+                    let backoff = config.control_backoff.clone();
+                    move |_| Ok(backoff.stream())
+                }))
+                .push(http_metrics::layer::<_, classify::Response>(
+                    ctl_http_metrics.clone(),
+                ))
+                .push(proxy::grpc::req_body_as_payload::layer().per_make())
+                .push(control::add_origin::layer())
+                .push_buffer_pending(
+                    config.destination_buffer_capacity,
+                    config.control_dispatch_timeout,
+                )
+                .into_inner()
+                .make(config.destination_addr.clone())
+        };
+
+        let resolver = api_resolve::Resolve::new(dst_svc.clone())
+            .with_context_token(&config.destination_context);
 
         let (tap_layer, tap_grpc, tap_daemon) = tap::new();
 
         // Spawn a separate thread to handle the admin stuff.
         {
             let (tx, admin_shutdown_signal) = futures::sync::oneshot::channel::<()>();
-            let local_identity = local_identity.clone();
+            let listen_identity = listen_identity.clone();
             let drain_rx = drain_rx.clone();
             thread::Builder::new()
                 .name("admin".into())
@@ -332,31 +345,33 @@ where
                     Runtime::new()
                         .expect("admin runtime")
                         .block_on(future::lazy(move || {
-                            info_span!("admin", listen.addr = %admin_listen.listen_addr())
+                            info_span!("admin", listen_addr=%admin_listener.listen_addr())
                                 .in_scope(|| {
                                     trace!("spawning");
                                     serve::spawn(
-                                        admin_listen,
+                                        admin_listener,
                                         tls::AcceptTls::new(
-                                            local_identity.clone(),
-                                            Admin::new(report, readiness, trace_handle)
+                                            listen_identity.clone(),
+                                            Admin::new(report, readiness, trace_level)
                                                 .into_accept(),
                                         ),
                                         drain_rx.clone(),
                                     );
                                 });
 
-                            if let Some((listen, config)) = tap {
-                                info_span!("tap", listen.addr = %listen.listen_addr()).in_scope(
+                            if let Some((listener, tap_svc_name)) = control_listener {
+                                info_span!("tap", listen_addr=%listener.listen_addr()).in_scope(
                                     || {
                                         trace!("spawning");
                                         tokio::spawn(tap_daemon.map_err(|_| ()).in_current_span());
                                         serve::spawn(
-                                            listen,
+                                            listener,
                                             tls::AcceptTls::new(
-                                                local_identity,
+                                                listen_identity,
                                                 tap::AcceptPermittedClients::new(
-                                                    config.permitted_peer_identities,
+                                                    std::sync::Arc::new(
+                                                        Some(tap_svc_name).into_iter().collect(),
+                                                    ),
                                                     tap_grpc,
                                                 ),
                                             ),
@@ -370,7 +385,7 @@ where
 
                             info_span!("dns").in_scope(|| {
                                 trace!("spawning");
-                                tokio::spawn(dns_task);
+                                tokio::spawn(dns_bg);
                             });
 
                             if let Some(d) = identity_daemon {
@@ -408,13 +423,21 @@ where
 
         // Build the outbound and inbound proxies using the dst_svc client.
 
-        let profiles_client =
-            ProfilesClient::new(dst_svc, Duration::from_secs(3), dst.context.clone());
+        let profiles_client = ProfilesClient::new(
+            dst_svc,
+            Duration::from_secs(3),
+            config.destination_context.clone(),
+        );
 
-        let trace_collector = trace_collector.map(|config| {
-            let svc = svc::stack(connect::svc(config.control.connect.keepalive))
-                .push(tls::client::layer(local_identity.clone()))
-                .push_timeout(config.control.connect.timeout)
+        let trace_collector_svc = config.trace_collector_addr.as_ref().map(|addr| {
+            let keepalive = if addr.addr.is_loopback() {
+                config.inbound_connect_keepalive
+            } else {
+                config.outbound_connect_keepalive
+            };
+            svc::stack(connect::svc(keepalive))
+                .push(tls::client::layer(listen_identity.clone()))
+                .push_timeout(config.control_connect_timeout)
                 // TODO: perhaps rename from "control" to "grpc"
                 .push(control::client::layer())
                 .push(control::resolve::layer(dns_resolver.clone()))
@@ -422,82 +445,79 @@ where
                 // HTTP metrics aren't useful for a client where we never read
                 // the response.
                 .push(reconnect::layer({
-                    let backoff = config.control.connect.backoff;
+                    let backoff = config.control_backoff.clone();
                     move |_| Ok(backoff.stream())
                 }))
                 .push(proxy::grpc::req_body_as_payload::layer().per_make())
                 .push(control::add_origin::layer())
                 .push_buffer_pending(
-                    config.control.buffer.max_in_flight,
-                    config.control.buffer.dispatch_timeout,
+                    config.destination_buffer_capacity,
+                    config.control_dispatch_timeout,
                 )
                 .into_inner()
-                .make(config.control.addr);
+                .make(addr.clone())
+        });
 
+        let spans_tx = trace_collector_svc.map(|trace_collector| {
             let (spans_tx, spans_rx) = mpsc::channel(100);
 
-            tokio::spawn({
-                let node = oc::Node {
-                    identifier: Some(oc::ProcessIdentifier {
-                        host_name: config.hostname.clone().unwrap_or_default(),
-                        pid: 0,
-                        start_timestamp: Some(start_time.into()),
-                    }),
-                    service_info: Some(oc::ServiceInfo {
-                        name: "linkerd-proxy".to_string(),
-                    }),
-                    ..oc::Node::default()
-                };
-                SpanExporter::new(svc, node, spans_rx, span_metrics)
-                    .map_err(|error| {
-                        error!(%error, "span exporter failed");
+            let node = oc::Node {
+                identifier: Some(oc::ProcessIdentifier {
+                    host_name: config.hostname.clone().unwrap_or_default(),
+                    pid: 0,
+                    start_timestamp: Some(start_time.into()),
+                }),
+                service_info: Some(oc::ServiceInfo {
+                    name: "linkerd-proxy".to_string(),
+                }),
+                ..oc::Node::default()
+            };
+            let span_exporter = SpanExporter::new(trace_collector, node, spans_rx, span_metrics);
+            tokio::spawn(
+                span_exporter
+                    .map_err(|e| {
+                        error!("span exporter failed: {}", e);
                     })
-                    .instrument(info_span!("opencensus-exporter"))
-            });
-
+                    .instrument(info_span!("opencensus-exporter")),
+            );
             spans_tx
         });
 
-        info_span!("out", listen.addr = %outbound_listen.listen_addr()).in_scope({
-            let dst = dst.clone();
-            || {
-                outbound::spawn(
-                    outbound_listen,
-                    outbound,
-                    local_identity.clone(),
-                    outbound::resolve(
-                        dst.get_suffixes,
-                        dst.control.connect.backoff.clone(),
-                        resolver,
-                    ),
-                    dns_resolver,
-                    profiles_client.clone(),
-                    dst.profile_suffixes.clone(),
-                    tap_layer.clone(),
-                    trace_collector.clone(),
-                    outbound_handle_time,
-                    endpoint_http_metrics.clone(),
-                    route_http_metrics.clone(),
-                    retry_http_metrics,
-                    transport_metrics.clone(),
-                    drain_rx.clone(),
-                )
-            }
+        info_span!("out", listen_addr=%outbound_listener.listen_addr()).in_scope(|| {
+            outbound::spawn(
+                &config,
+                listen_identity.clone(),
+                outbound_listener,
+                outbound::resolve(
+                    config.destination_get_suffixes.clone(),
+                    config.control_backoff.clone(),
+                    resolver,
+                ),
+                dns_resolver,
+                profiles_client.clone(),
+                tap_layer.clone(),
+                outbound_handle_time,
+                endpoint_http_metrics.clone(),
+                route_http_metrics.clone(),
+                retry_http_metrics,
+                transport_metrics.clone(),
+                spans_tx.clone(),
+                drain_rx.clone(),
+            )
         });
 
-        info_span!("in", listen.addr = %inbound_listen.listen_addr()).in_scope(move || {
+        info_span!("in", listen_addr=%inbound_listener.listen_addr()).in_scope(move || {
             inbound::spawn(
-                inbound_listen,
-                inbound,
-                local_identity,
+                &config,
+                listen_identity,
+                inbound_listener,
                 profiles_client,
-                dst.profile_suffixes,
                 tap_layer,
                 inbound_handle_time,
                 endpoint_http_metrics,
                 route_http_metrics,
                 transport_metrics,
-                trace_collector,
+                spans_tx,
                 drain_rx,
             )
         });
